+++ axon/HTTP_transport/client.py
@@ -1,0 +1,19 @@
+from sys import path
+path.append('..')
+
+from axon.serializers import serialize, deserialize
+from axon.transport_client import http, req_executor, error_handler, AsyncResultHandle
+
+class HTTPTransportClient():
+
+	def __init__(self):
+		pass
+
+	def call_rpc_helper(self, url, data):
+		resp = http.request('POST', url, fields=data)
+		result_str = error_handler(resp.data.decode())
+		return deserialize(result_str)
+
+	def call_rpc(self, url, args, kwargs):
+		future = req_executor.submit(self.call_rpc_helper, url, {'msg': serialize((args, kwargs))})
+		return AsyncResultHandle(future)
