+++ axon/HTTP_transport/config.py
@@ -1,0 +1,2 @@
+
+scheme = 'http'
