+++ axon/config.py
@@ -1,10 +1,10 @@
 from types import SimpleNamespace
-from .inline_executor import InlineExecutor
+from axon.inline_executor import InlineExecutor
 
-from .transport_worker import HTTPTransportWorker
-from .transport_client import HTTPTransportClient
-from .socket_worker import SocketTransportWorker
-from .socket_client import SocketTransportClient
+from axon.socket_worker import SocketTransportWorker
+from axon.socket_client import SocketTransportClient
+
+import axon.HTTP_transport as transport
 
 inline_executor = InlineExecutor()
 
@@ -12,10 +12,8 @@ version = "0.2.1"
 default_service_depth = 3
 default_rpc_endpoint = 'rpc'
 NUM_OPEN_REQS = 512
-default_client_tl = HTTPTransportClient()
-url_scheme = 'http'
-# default_client_tl = SocketTransportClient()
-# url_scheme = 'ws'
+default_client_tl = transport.client()
+url_scheme = transport.config.scheme
 
 comms_config = {
 	'notice_board_port': 8002,
@@ -28,6 +26,5 @@ comms_config = SimpleNamespace(**comms_config)
 default_service_config = {
 	'endpoint_prefix': '',
 	'executor': inline_executor,
-	'tl': HTTPTransportWorker(comms_config.worker_port)
-	# 'tl': SocketTransportWorker(comms_config.worker_port)
+	'tl': transport.worker(comms_config.worker_port)
 }
