+++ axon/tests/notice_board_test.py
@@ -1,8 +1,6 @@
 import time
 from threading import Thread
 
-import sys
-sys.path.append('..')
 import axon
 
 # the notice board will be run in a separate thread so we can run a client in the main thread to test it
