+++ axon/tests/service_client_test.py
@@ -1,6 +1,3 @@
-from sys import path
-path.append('..')
-
 import axon
 import asyncio
 import time
